Colony v5.22

+++ SG14_test/plf_colony_test.cpp
@@ -932,8 +932,8 @@ void plf_colony_test()
 					size = static_cast<unsigned int>(i_colony.size());
 					range1 = xor_rand() % size;
 					range2 = range1 + 1 + (xor_rand() % (size - range1));
-					i_colony.advance(it1, range1);
-					i_colony.advance(it2, range2);
+					i_colony.advance(it1, static_cast<int>(range1));
+					i_colony.advance(it2, static_cast<int>(range2));
 
 					i_colony.erase(it1, it2);
 
@@ -986,8 +986,8 @@ void plf_colony_test()
 					size = static_cast<unsigned int>(i_colony.size());
 					range1 = xor_rand() % size;
 					range2 = range1 + 1 + (xor_rand() % (size - range1));
-					i_colony.advance(it1, range1);
-					i_colony.advance(it2, range2);
+					i_colony.advance(it1, static_cast<int>(range1));
+					i_colony.advance(it2, static_cast<int>(range2));
 
 					i_colony.erase(it1, it2);
 
@@ -1168,8 +1168,8 @@ void plf_colony_test()
 					size = static_cast<unsigned int>(ss_nt.size());
 					range1 = xor_rand() % size;
 					range2 = range1 + 1 + (xor_rand() % (size - range1));
-					ss_nt.advance(ss_it1, range1);
-					ss_nt.advance(ss_it2, range2);
+					ss_nt.advance(ss_it1, static_cast<int>(range1));
+					ss_nt.advance(ss_it2, static_cast<int>(range2));
 
 					ss_nt.erase(ss_it1, ss_it2);
